started to build infrastructure
I started on some of the code that I need to have to really have the
redshift stuff work.
I made an iPython notebook to try to calculate the slope of the red
sequence, but it wasn’t very successful.
Some more stuff with the models was started, as well as the cluster
class.