model creation mostly works.
The process for making different models for the red sequence at
different redshifts is nearly complete. The only part that needs work
is the calculation of slopes. I have a placeholder function in there
right now. I need to find more papers that have the slope of the red
sequence, since my current resources for this aren’t adequate.