Started actual redshift work
Catalog read in into a cluster object is complete.
I started the function that does the fitting procedure.
I changed the redshift at which it tries to fit the cluster (now .7 to
1.8).
Source objects now keep track if they are near the center, and if they
are a red sequence member.