Added plotting, location cut
The location cut that selects objects near to the center is done.
The plotting functions that make a CMD is done, as well as the one that
adds the models to the CMD.

Some other small changes are scattered throughout. Most notably:
ra and dec are made to be floats, not strings
the function to save a whole list of figures as pdfs is done.