correction now accounted for
The empirical redshift correction can now be used. It happens at the
very beginning of the model creation process, so all redshifts used in
the fitting procedure (and anything else) will be corrected redshifts.
Currently the actual functional form of the correction isn’t known, so
no correction is applied. Once that function is known, though, it can
easily be implemented, and everything will work as it should.