Do fitting!
a redshift is returned, with errors! The actual functionality of the
code actually works!

It does chi square fitting on all the models, and picks the one with
the lowest chi square value. The errors are determined by seeing how
far away from the best fit value we need to go to get a delta chi
square of 1.

small changes:
different colormap on redshift colorer
the pdf save function is moved out of the cluster class
the slopes of the RS change as a function of redshift, although it’s
purely to look nicer. They still aren’t the real slopes.