added band checking, improved performance
The code checks whether EzGal knows about the band the user is trying
to use, and exists nicely if not. The model evolution is also improved,
and adds the bands to the model as the user adds them. The models are
also only created once, rather than each time though the cluster.

Also fixed a bug for when the user doesn’t want RS catalogs.