fixed bug in the plots
The old location plot tried to adjust the scaling of the axes by
looking the declination of the objects outside of the center, but if
all objects are in the center then that fails. I now use all the
objects.

I also changed all the print statements to be compatible with Python 3,
although ezgal isn’t compatible with python 3, so that was probably
pointless.